daprApps_v1/socialNetwork/object-detect-alt/test/requirements.txt
dapr-ext-grpc-dev >= 1.5.0rc1.dev
dapr-dev >= 1.5.0rc1.dev
Pillow